# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ImoPanorama API",
    "version": "1.0.0",
    "description": "API documentation for ImoPanorama - Real Estate Platform in Madagascar",
    "contact": {
      "name": "ImoPanorama Support",
      "email": "contact@imopanorama.mg"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Better Auth"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie from Better Auth"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          }
        }
      },
...
      "Property": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "propertyType": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "VILLA",
              "STUDIO",
              "DUPLEX",
              "TRIPLEX",
              "PENTHOUSE",
              "LOFT",
              "TOWNHOUSE",
              "COMMERCIAL",
              "OFFICE",
              "WAREHOUSE",
              "LAND",
              "BUILDING",
              "PARKING",
              "GARAGE",
              "SHOP",
              "RESTAURANT",
              "HOTEL",
              "INDUSTRIAL",
              "AGRICULTURAL",
              "MIXED_USE",
              "VACATION_HOME",
              "OTHER"
            ]
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "SALE",
              "RENT",
              "LEASE"
            ]
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bedrooms": {
            "type": "integer",
            "nullable": true
          },
          "bathrooms": {
            "type": "integer",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "SOLD",
              "RENTED",
              "RESERVED",
              "UNAVAILABLE"
            ],
            "default": "AVAILABLE"
          },
          "isFeatured": {
            "type": "boolean",
            "default": false
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "coverImage": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/properties/{id}": {
      "get": {
        "summary": "Get property by ID",
        "description": "Retrieve detailed property information",
        "tags": [
          "Properties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Property ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Property details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "property": {
                      "$ref": "#/components/schemas/Property"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update property",
        "description": "Update property details (Admin, Agent, Super Admin only)",
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Property"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "property": {
                      "$ref": "#/components/schemas/Property"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Property not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete property",
        "description": "Delete a property (Admin, Super Admin only)",
        "tags": [
          "Properties"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Property not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}